Hey Erick,
I added a couple signal paths to the motion controller.  They are labeled "Computed Torque Command Feedforward".  By using estimates of the mass and damping in the feedforward path, the best estimates of those parameters will be achieved.  I can discuss some experimental procedure concerning how to tune the mass and damping estimates when the time comes.

Also,  when we met with David we saw that the force command was jittering because the system was operating at no-load.  This might actually happen on the physical system when you implement control.  After thinking about this a little bit, I think there is a combination of sources leading to this jitter: the actuator can make pretty huge forces and has a huge back-EMF, the print head's equivalent mass through the gear train is extremely small, and the sample period is low.

My best suggestion for you guys is to add physical mass to the actuator's print head (same concept as a flywheel).  That way the inertial forces will be larger and jitter will be significantly reduced.  I did this in the model by multiplying the physical mass by 100.  You can play with it and see how this will significantly improve the no-load behavior.

Thanks,
Caleb