Updated AutomatedTransfer_v2 and accompanying VIs (folder: crio-FPGA-control-testing-v5), now with: 1. All of the new variables kvc, ymod, mar, etc. for the updated control law are passed down into 'FPGA Main'. The variables kvc, ymod, mar, bp, ba, ksa, kisa, k_backemf, k_1, K_2, Frr, xrr, Foff, and Fthresh are passed via the cluster 'Digital Controller'. The variables bo, kso, and kiso are passed via the cluster 'Motion Observer'. 2. The clusters 'Digital Controller' and 'Motion Observer' are wired into 'R2R-Controller.vi'.
Notes:
i. The word lengths for the new variables (all fixed point) need to be changed based on the expected ranges and required precisions of the numbers. The new clusters (Digital Controller and Motion Observer) are typedefs, so those changes can be made to the typedefs.
ii. Only 3 variables from 'Digital Controller' (kvc, Foff, and Fthresh) are being used in the present control law.